refactor(map): extract role glyph SVG builders into Leaflet-free module
Move repeaterTowerSvg, roleGlyphInnerSvg, roleGlyphMarkerSvg,
unmessageableBadgeSvg, and getHopColor out of markerIcons.ts into the
new src/utils/roleGlyphSvg.ts verbatim (byte-identical output), so the
upcoming Node Details traceroute strip (#4381 WP1) can import them
without pulling Leaflet into a non-map bundle/test. markerIcons.ts
re-exports everything so no existing importer changes. Also exports
unmessageableBadgeSvg (was module-private) and adds plainNodeDiscSvg
and unknownNodeSvg for the strip's fallback/unknown-hop glyphs.

Extends markerIcons.test.ts with an extraction guard asserting the
re-exported functions are the same references (and produce identical
output) as the ones in roleGlyphSvg.ts, plus a Leaflet/React-import
check on the new module.
